Created the pullable block. Saved the test scene as Level1.
CobblestoneFloor2.png: added to be used by the draggable block.

DraggableBlock: added to allow the Gameobject to be dragged horizontally within a certain range by anything calling its Move function. The Move function just takes the speed as is, so the caller needs to make sure the speed is correct for the amount of times they're called.

Level1: added as a test scene for now, but will be used for the first level of the game.

PullableWall: added to be used in multiple places in a level, can be dragged horizontally within a given range.