Added blocks that can be moved verticaly, like the previous PullableWal.
DraggableBlock: now inherits from the MovableBlock abstract class. Kept a test Update function in.

Level1: added a WeighableWall for testing.

MovableBlock: added to be the base class of any script that allows the attached gameobject to be moved in a linear line in the given range.

WeighableBlock: added to allow hanging blocks to be pulled down. Inherits from MoveableBlock. Has a test Update function defined.

WeighableWall.prefab: can be pulled down when its Move function is called.